# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: SLIDE Backend API
  description: |
    DEALT/SLIDE game server. Flask is the authoritative game server; Supabase provides DB/Auth/Realtime.
    All endpoints require `Authorization: Bearer <JWT>` unless marked as public.
    In dev mode (no SUPABASE_URL configured) a mock DEV_USER is injected automatically.
  version: "1.0.0"
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.slide.game
    description: Production (placeholder)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Coordinates:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
      required: [lat, lng]
    AddressResult:
      type: object
      properties:
        address:
          type: string
        formattedAddress:
          type: string
        coordinates:
          $ref: "#/components/schemas/Coordinates"
        city:
          type: string
        inServiceArea:
          type: boolean
        mapboxId:
          type: string
    Block:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        formattedAddress:
          type: string
        city:
          type: string
        coordinates:
          $ref: "#/components/schemas/Coordinates"
        ownerId:
          type: string
        gangName:
          type: string
        gridData:
          type: object
        trafficScore:
          type: number
        incomePerHour:
          type: number
        heatLevel:
          type: number
        claimedAt:
          type: string
          format: date-time
        backgrounds:
          $ref: "#/components/schemas/BlockBackgrounds"
    BlockBackgrounds:
      type: object
      properties:
        topdownUrl:
          type: string
          description: Mapbox static satellite image URL
        streetNUrl:
          type: string
          description: Google Street View north heading
        streetEUrl:
          type: string
          description: Google Street View east heading
        streetSUrl:
          type: string
          description: Google Street View south heading
        streetWUrl:
          type: string
          description: Google Street View west heading
...
security:
  - bearerAuth: []
paths:
...
  /api/blocks/search:
    get:
      summary: Search for addresses
      security: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query (min 3 chars)
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Address search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/AddressResult"
                  query:
                    type: string
  /api/blocks/claim:
    post:
      summary: Claim a block for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
                coordinates:
                  $ref: "#/components/schemas/Coordinates"
                city:
                  type: string
                  enum: [nyc, la, miami, chicago, detroit, nola]
                gangName:
                  type: string
              required: [address, coordinates, city]
      responses:
        "201":
          description: Block claimed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        "400":
          description: Validation error or unsupported city
        "409":
          description: Block already claimed
  /api/blocks/{block_id}:
    get:
      summary: Get block by ID
      security: []
      parameters:
        - name: block_id
          in: path
          required: true
          schema:
            type: string
        - name: includeGrid
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Block details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        "404":
          description: Block not found
  /api/blocks/{block_id}/generate_backgrounds:
    post:
      summary: Generate satellite and street-view backgrounds for a block
      parameters:
        - name: block_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Backgrounds generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  block_id:
                    type: string
                  backgrounds:
                    $ref: "#/components/schemas/BlockBackgrounds"
                  anchors_json:
                    type: object
                    description: 8x8 grid anchor lat/lon mapping
...
  /api/blocks/city/{city}:
    get:
      summary: Get blocks in a city
      security: []
      parameters:
        - name: city
          in: path
          required: true
          schema:
            type: string
            enum: [nyc, la, miami, chicago, detroit, nola]
      responses:
        "200":
          description: City blocks
          content:
            application/json:
              schema:
                type: object
                properties:
                  blocks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Block"
                  count:
                    type: integer
...